[tests] Split mscorlib into two test assemblies. Fixes #41746.
On watchOS we can't include all the tests for device in a single
assembly, because the native object code becomes too big.

So instead split the tests into two assemblies. Luckily this
is fairly easy for test projects, since they have few dependencies
between source files (there are some however, so source files
are grouped per directory so that files in the same directory do
not end up in different test projects).

https://bugzilla.xamarin.com/show_bug.cgi?id=41746