[xharness] Make the top-level test names identical between iOS and Mac.
Make the top-level test name for equivalent tests between iOS and Mac have the
same name, so that these tests are grouped together.

For instance: we used to have a System-Mac tests and System tests (the latter
being iOS tests), now there are just System tests, and nested inside are both
the Mac and iOS System tests.