DROP [tests] Don't put entire cmd output in exception message.
[mtouch tests] Refactor registrar tests a little bit to not cause trouble in Xamarin Studio.

Change how the registrar calls mtouch a bit, so that we don't use API that
throws an exception with the entire output in the exception message.

The problem is that if the process has a lot of output, XS slows down to a
crawl when showing the exception message in the test result pad.

Instead print the output to the terminal.