[mtouch/mmp] Try to fix version creation.
The following are not identical:

    new Version (1, 2);

and

    new Version (1, 2, 0)

what's even worse is that v1.2.0 > v1.2 (!), which means that we can't pass 0
as the default value for a version field.

Instead rewrite the overloads to create the Version instance using the correct
constructor overload.

Hopefully this fixes the following confusing error:

error MT4134: Your application is using the 'Intents' framework, which isn't included in the watchOS SDK you're using to build your app (this framework was introduced in watchOS 3.2.0, while you're building with the watchOS 3.2 SDK.) [...]