MISSING TEST [mtouch/mmp] When gathering frameworks, gather into the product assembly's frameworks.
When gathering frameworks, gather into the product assembly's frameworks, not
the global frameworks.

This is necessary when building a non-linked dylib/framework for
Xamarin.iOS.dll, since otherwise we won't know how to link with the right
frameworks.

This went unnoticed for a while, because we also process P/Invokes to find
frameworks to link with, and most frameworks (but not CoreImage) have
corresponding P/Invokes.