[link sdk] Modify the test for bug #56876 so that it doesn't trigger MT4169.
This test looks strange, but it was how I could trigger the invalid AOT-
compilation output found in bug #56876. The important part is the parameter to
the P/Invoke, but unfortunately we run into problems with such a parameter in
P/Invokes to objc_msgSend when generating P/Invoke wrappers for objc_msgSend
(i.e. when building the linksdk tests for watchOS/release):

> MT4168: Failed to generate a P/Invoke wrapper for objc_msgSend(LinkSdk.Aot.AotBugsTest.сolor_bug_56876): The registrar cannot get the ObjectiveC type for managed type `LinkSdk.Aot.AotBugsTest.сolor_bug_56876`

Fortunately the fix is trivial: use a different native function (I picked
sel_registerName, but it doesn't really matter as long as it's not in the
objc_msgSend family of functions) to avoid the P/Invoke wrapper generation
(which only occurs for the objc_msgSend family of functions).