[mtouch] Show MT0091 as a warning if we're using the dynamic registrar.
Even if using an Xcode older than the bindings we're shipping, everything
might just happen to work when using the dynamic registrar, so downgrade the
MT0091 error to a warning in that case (and cross fingers, because things
might still break).

Hopefully this makes it easier when using older Xcodes, because the typical
scenario of Sim+DontLink and Dev+LinkSdk/LinkAll will now just work (if the
app developer also crosses fingers piously and sacrifies the appropriate
amount of pizza to the proper deity).

Reference: https://trello.com/c/wP3Pcb1q/692-targeting-older-xcodes-without-enabling-managed-linking