[monotouch-test] The media library permission API is only available in iOS 9.3+. Fixes #59995.
The media library permission API (to either query the permission status or ask
for permission) was introduced in iOS 9.3, so we need to make sure to not call
it on earlier iOS versions.

If running on older iOS versions, then just don't do anything (unless we're
asked to ignore tests that put up permission dialogs, in which case ignore the
test, since that's the safe approach).

https://bugzilla.xamarin.com/show_bug.cgi?id=59995