[ObjCRuntime] Add a function that can compare a managed string to a native UTF8 string without memory allocations.
Add a function that can compare a managed string to a native UTF8 string
without memory allocations for the common scenario without multibyte UTF8
characters.

If there are multibyte UTF8 characters, the code falls back to the simple
behavior of converting the native UTF8 string to a managed string before
comparing.