[src] Fix rebuilding pdb files.
There was a problem with the logic to install pdb files, where they'd only be
touched if they already existed instead of updated.

This causes frequent problems when building locally, because symbols would be
out of date (resulting in build warnings, broken debugging, etc).

Since it only happens when assemblies are rebuilt, this was not a problem for
packages built on wrench.