[mk] Put shell code in shell scripts and add support for switching between different remotes for dependencies.
Put the shell code for resetting README dependencies in a shell script instead
of embedded in the Makefile so that it's easier to write, read and debug.

Also add support for switching between different remotes for dependencies,
automatically computing the remote from the GitHub url (a side effect is that
all README dependencies will now end up with a 'xamarin' remote in addition to
the 'origin' remote, but this should have no other effect).