[mtouch] Unify code to detect and handle frameworks that aren't supported in the simulator. Fixes #4422.
Unify the code to detect frameworks that aren't supported in the simulator (we
had switches in two places, now this data is stored per framework).

Also remove some of the Metal frameworks for some of the platforms from these
lists (since they're now available in the simulator in some cases), which
fixes #4422.

https://github.com/xamarin/xamarin-macios/issues/4422