[mtouch] Pass outfile= to the AOT compiler when compiling to full llvm as well.
The AOT compiler uses the 'outfile' as the base for a temporary .bc file. If
it's not set, the path to the assembly is used as the base instead. This does
not work if we compile the same assembly to multiple architectures
(armv7k+arm64_32), because the temporary file will clash between those AOT
compilations.