[xharness] Don't group simulators when running in server mode.
Grouping simulators means we have to wait until we know the simulators that
are available, because we group by the simulator's UDID.

When running in server mode, we don't need to group simulator test runs,
because we run the as the user chooses, not in any particular order.

This way the test list loads faster at startup, since we don't have to wait
until we've loaded the simulators before we can show the list.