[xharness] Use Cecil to inspect assemblies, instead of Reflection.
This is slightly faster (probably because reflection tries to load a lot of
other referenced assemblies, which may or may not exist, causing exceptions
(if they don't exist) or spend time loading them (which Cecil won't)).

It also avoids a lot of exception details showing up when tracing xharness
execution.