[mtouch/mmp] Move Application.BuildTarget to shared code.
Also add a 'None' build target for when we're building for neither simulator
nor device (i.e. macOS). This means the default value will change (since
'Simulator' is no longer the first value), but as far as I can tell we're
always assigning a specific value and not relying on the default, so this
should not make any difference.

This will be needed when the .NET code starts using these classes.