[tests] Remove the mono-native CheckSymbols test.
It's running into different behavior between OS versions, and it's getting
time-consuming and annoying to keep up with the changes.

So instead just remove the test, because as far as I can see it's not testing
something important anyway (we don't care exactly which library provides the
'mono_native_initialize' function, the important part is that it's callable,
which we already test elsewhere).