[mmp] Always set DEVELOPER_DIR when executing executables from Apple's toolchain. Fixes #8133.
This fixes strange errors if there's no Xcode configured in Visual Studio, and
the Xcode we're building with isn't the system Xcode. If this happened, any of
Xcode's tools would claim that the tool didn't exist, and that the user would
have to install the command-line tools (which might already be installed).

Fixes https://github.com/xamarin/xamarin-macios/issues/8133.