[CoreFoundation] Fix CFBundle.Architecture constants.
There was a confusion between hex and decimal values, the values are written
using decimal in the headers, but were treated as hex values in the C# source.

This is a breaking change, but the values are just plain wrong, so I don't
really see a way around this. Also there's no good way to work around it by
accepting the old values and converting in our managed wrapper method, because
the old ARM values match the new PPC values, so there's no way to know if the
input is using old values or new values.