[msbuild] Rework app manifest properties a bit.
* _InputAppManifest: this is the input Info.plist (typically just
  'Info.plist', from the project directory).
* _AppManifest: this is the Info.plist in the app bundle, fully compiled.

This way any task that takes the '_AppManifest' variable, will now read the
compiled Info.plist, and not the input Info.plist. This is important, because
the input Info.plist will often not contain all the information (there are
MSBuild variables that can end up in the final Info.plist, as well as other
partial plists that can be merged into it).

Also change the CompileAppManifest task to take the full path to the compiled
app manifest, to simplify code a bit.