[msbuild] Rework app manifest compilation a bit.
The CompileAppManifest task is now entirely responsible for compiling the app
manifest, and computing everything related with it (implemented in previous
commits).

The new ReadAppManifest task will read the compiled app manifest, and populate
a few MSBuild variables accordingly. This is not done in the
CompileAppManifest target, because we'll need these variables in every build,
but the CompileAppManifest target might not run if its output is not out of
date.

This means that:

* The DetectSigningIdentity now depends on CompileAppManifest (instead of the
  other way around, when DetectSigningIdentity would compute the bundle
  identifier - which is now CompileAppManifest's responsibility).
* A lot of other targets must depend on ReadAppManifest, so that the MSBuild
  variables they need are populated.
* To clean up the code a bit, several DependsOnTargets target lists were
  separated out to *DependsOn properties.