[msbuild] Rework how app manifest detection works.
Our current behavior is to detect any None, BundleResource or Content item
that's named 'Info.plist', and assume that's the app manifest.

That doesn't quite work when we end up with multiple 'Info.plist' entries in
any of those item groups (one example being a framework as a BundleResource -
all frameworks have an Info.plist, and there's no good way to distinguish what
the developer's intention was).

So:

1. Implement a 'AppManifestDetectionEnabled' property to disable automatic app manifest detection.
2. Add a public 'AppBundleManifest' property that specifies the app manifest (this is just a renamed version of our previously private '_AppManifest' property).

This makes it possible for app developers to:

* Disable automatic app manifest detection.
* Still have an app manifest by specifying it manually.

Also:

* Rename '_AppBundleManifest' '_AppBundleManifestPath' to make it less
  confusing with the new 'AppBundleManifest' property.