[msbuild] Modify the computation of logical paths for resources for .NET.
REVERT THIS IT IS WRONG - FIGURE OUT SOMETHING ELSE.

The legacy behavior is to use the project where an item is defined to compute
the logical path. This works just as expected if the item is defined in the
project file itself (and the behavior won't change).

However, if the item is included from another imported project file, there's a difference.

Given this file hierarchy:

    Resources.csproj
    Dir/Project.csproj
    Dir/Resource.png

and these are the files:

Resources.csproj:

```xml
<ItemGroup>
    <Content Include="Resource.png" />
</ItemGroup>
```

and Dir/Project.csproj:

```xml
<Import Project="../Resources.csproj" />
```

then the defining project of Resource.png is "Resources.csproj", and the
logical path to the Resource.png file is "Dir/Resources.png".

This was implemented quite a while ago (in maccore [2]), but I don't think
this was the intended behavior, and it seems quite unintuitive, especially
since the paths used in the Include attributes are resolved based on the main
project file, not any additional imported files:

    The Include attribute is a path that is interpreted relative to the
    project file's folder, $(MSBuildProjectPath), even if the item is in an
    imported file such as a .targets file. [1]

The intention here is to always compute the logical path relative to the main
project, just like other paths are resolved for items.

In order to avoid breaking legacy Xamarin projects, the new behavior is only
implemented for .NET.

FIXME: add tests

[1]: https://docs.microsoft.com/en-us/visualstudio/msbuild/msbuild-items?view=vs-2019#create-items-in-a-project-file
[2]: https://github.com/xamarin/maccore/commit/2d8d039e512f8bbd0624320fd7863bfccda7f5cf