[src] Add a new ObjCRuntime.NativeHandle type to represent native handles for .NET. Fixes #13126.
Add a new struct, ObjCRuntime.NativeHandle, which will be used to represent native
handles for .NET (instead of using IntPtr). The main purpose is to be able to use
'nint' as a number in API while not being prevented from using native handles as
well.

One example is NSMutableString, which has a constructor that takes a single 'nint
capacity' parameter. With this change, we'll also be able to have a constructor that
takes a native handle in .NET - otherwise we'd have two constructors with the same
signature, because a C# 'nint' is just an 'IntPtr'.

Fixes https://github.com/xamarin/xamarin-macios/issues/13126.